refactor(type): split execution overlay types
