+++ Docs/01_Work_List/W05_Code_Quality_Plan/W05_Type_Header_Organization_Work_Plan.md
@@ -32,7 +32,7 @@ refactor/type-header-organization
 - [x] hit / damage / combat result 타입 분리
 - [x] combat signal source / target 타입 분리
 - [x] action / reaction 저장형 data 타입 분리
-- [ ] execution / observable overlay 타입 분리
+- [x] execution / observable overlay 타입 분리
 - [ ] orchestration / feedback 타입 정리
 - [ ] 사용처 include 교체 및 umbrella 의존 제거
 - [ ] 최종 빌드 / PIE / 로그 검증
