+++ Source/Portfolio/Type/CObservableOverlayTypes.cpp
@@ -1,0 +1,1 @@
+#include "Type/CObservableOverlayTypes.h"
