+++ Source/Portfolio/Type/CWeaponStructure.h
@@ -7,6 +7,8 @@
 #include "Type/CExecutionRuleTypes.h"
 #include "Type/CActionDataTypes.h"
 #include "Type/CReactionDataTypes.h"
+#include "Type/CObservableOverlayTypes.h"
+#include "Type/CExecutionTypes.h"
 #include "Type/CCombatHitTypes.h"
 #include "Type/CCombatDamageTypes.h"
 #include "Type/CCombatResultTypes.h"
@@ -15,134 +17,6 @@
 #include "Type/CStateTypes.h"
 #include "Type/CHealthTypes.h"
 #include "CWeaponStructure.generated.h"
-
-UENUM(BlueprintType)
-enum class EExecutionDecision : uint8
-{
-	None = 0,
-
-	Reject,
-	Ignore,
-
-	Accept,
-
-	Max,
-};
-
-UENUM(BlueprintType)
-enum class EExecutionRelationship : uint8
-{
-	None = 0,
-
-	Independent,
-	Sequential,
-	Exclusive,
-
-	Max,
-};
-
-UENUM(BlueprintType)
-enum class EExecutionApplyMode : uint8
-{
-	None = 0,
-
-	Start,
-	Reserve,
-	Intervene,
-
-	Max,
-};
-
-UENUM(BlueprintType)
-enum class EObservableOverlayHandling : uint8
-{
-	None = 0,
-
-	ClearGuardState,
-	ClearGuardOverlay,
-
-	Max,
-};
-
-UENUM(BlueprintType)
-enum class EExecutionStopReason : uint8
-{
-	None = 0,
-
-	Interrupted,
-	Ignored,
-
-	Max,
-};
-
-UENUM(BlueprintType)
-enum class EObservableOverlayEventType : uint8
-{
-	None = 0,
-
-	GuardInputPressed,
-	GuardInputReleased,
-
-	GuardInStarted,
-	GuardOutStarted,
-
-	SwitchToGuard,
-	AllowGuardStart,
-
-	GuardLifecycleCompleted,
-	GuardLifecycleInterrupted,
-
-	Max,
-};
-
-USTRUCT(BlueprintType)
-struct FObservableOverlayEventContext
-{
-	GENERATED_BODY()
-
-public:
-	UPROPERTY(Transient)
-	EObservableOverlayEventType EventType = EObservableOverlayEventType::None;
-
-public:
-	FObservableOverlayEventContext() = default;
-
-	explicit FObservableOverlayEventContext(EObservableOverlayEventType InEventType)
-		: EventType(InEventType)
-	{
-	}
-
-public:
-	bool IsValidMinimal() const
-	{
-		return EventType != EObservableOverlayEventType::None && EventType != EObservableOverlayEventType::Max;
-	}
-};
-
-UENUM(BlueprintType)
-enum class EExecutionStopSource : uint8
-{
-	None = 0,
-
-	ActionOrchestration,
-	ReactionOrchestration,
-
-	System,
-	External,
-
-	Max,
-};
-
-UENUM(BlueprintType)
-enum class EExecutionAfterStopAction : uint8
-{
-	None = 0,
-
-	StopOnly,
-	StartIncoming,
-
-	Max,
-};
 
 UENUM(BlueprintType)
 enum class EActionStopSource : uint8
@@ -358,287 +232,6 @@ enum class EActionSFXPlayType : uint8
 {
 	Once,
 	Loop
-};
-
-USTRUCT(BlueprintType)
-struct FGuardObservableOverlaySnapshot
-{
-	GENERATED_BODY()
-
-public:
-	UPROPERTY(Transient)
-	bool bWantsGuarding = false;
-
-	UPROPERTY(Transient)
-	bool bIsGuardingPose = false;
-
-	UPROPERTY(Transient)
-	bool bCanGuard = false;
-
-	UPROPERTY(Transient)
-	bool bCanParry = false;
-
-	UPROPERTY(Transient)
-	bool bCanStartGuard = true;
-
-public:
-	bool HasGuardOverlay() const
-	{
-		return bIsGuardingPose || bCanGuard || bCanParry;
-	}
-
-	bool HasGuardRuntimeState() const
-	{
-		return !bCanStartGuard || bWantsGuarding || HasGuardOverlay();
-	}
-};
-
-USTRUCT(BlueprintType)
-struct FObservableOverlaySnapshot
-{
-	GENERATED_BODY()
-
-public:
-	UPROPERTY(Transient)
-	FGuardObservableOverlaySnapshot Guard = FGuardObservableOverlaySnapshot();
-
-public:
-	bool HasObservableOverlay() const
-	{
-		return Guard.HasGuardOverlay();
-	}
-};
-
-USTRUCT(BlueprintType)
-struct FExecutionSnapshot
-{
-	GENERATED_BODY()
-
-public:
-	UPROPERTY(Transient)
-	EExecutionState ExecutionState = EExecutionState::Idle;
-
-public:
-	UPROPERTY(Transient)
-	bool bIsDead = false;
-
-	UPROPERTY(Transient)
-	FObservableOverlaySnapshot ObservableOverlay = FObservableOverlaySnapshot();
-
-public:
-	bool IsIdle() const
-	{
-		return ExecutionState == EExecutionState::Idle;
-	}
-
-	bool IsInAction() const
-	{
-		return ExecutionState == EExecutionState::Action;
-	}
-
-	bool IsInReaction() const
-	{
-		return ExecutionState == EExecutionState::Reaction;
-	}
-
-	bool IsDead() const
-	{
-		return bIsDead || ExecutionState == EExecutionState::Dead;
-	}
-
-	bool HasObservableOverlay() const
-	{
-		return ObservableOverlay.HasObservableOverlay();
-	}
-};
-
-USTRUCT(BlueprintType)
-struct FExecutionParticipant
-{
-	GENERATED_BODY()
-
-public:
-	UPROPERTY(Transient)
-	bool bIsValid = false;
-
-	UPROPERTY(Transient)
-	EExecutionDomain ParticipantDomain = EExecutionDomain::None;
-
-	UPROPERTY(Transient)
-	FActionExecutionContext ActionContext = FActionExecutionContext();
-
-	UPROPERTY(Transient)
-	FReactionExecutionContext ReactionContext = FReactionExecutionContext();
-
-public:
-	bool IsValidMinimal() const;
-	bool IsActionParticipant() const;
-	bool IsReactionParticipant() const;
-
-	const FActionExecutionContext& GetActionContext() const;
-	const FReactionExecutionContext& GetReactionContext() const;
-
-	UObject* GetExecutor() const;
-	int32 GetPriority() const;
-};
-
-USTRUCT(BlueprintType)
-struct FObservableOverlayExecutionDecision
-{
-	GENERATED_BODY()
-
-public:
-	UPROPERTY(Transient)
-	EExecutionDecision Decision = EExecutionDecision::None;
-
-	UPROPERTY(Transient)
-	TArray<EObservableOverlayHandling> Handlings;
-
-public:
-	bool IsAccepted() const
-	{
-		return Decision == EExecutionDecision::Accept;
-	}
-};
-
-USTRUCT(BlueprintType)
-struct FExecutionDecisionQuery
-{
-	GENERATED_BODY()
-
-public:
-	UPROPERTY(Transient)
-	FExecutionSnapshot Snapshot = FExecutionSnapshot();
-
-	UPROPERTY(Transient)
-	FExecutionParticipant IncomingPart = FExecutionParticipant();
-
-	UPROPERTY(Transient)
-	FExecutionParticipant ActivePart = FExecutionParticipant();
-
-public:
-	bool HasIncomingPart() const
-	{
-		return IncomingPart.IsValidMinimal();
-	}
-
-	bool HasActivePart() const
-	{
-		return ActivePart.IsValidMinimal();
-	}
-};
-
-USTRUCT(BlueprintType)
-struct FObservableOverlayQuery
-{
-	GENERATED_BODY()
-
-public:
-	UPROPERTY(Transient)
-	FExecutionDecisionQuery DecisionQuery = FExecutionDecisionQuery();
-
-	UPROPERTY(Transient)
-	EExecutionApplyMode ApplyMode = EExecutionApplyMode::None;
-};
-
-USTRUCT(BlueprintType)
-struct FExecutionDecisionResult
-{
-	GENERATED_BODY()
-
-public:
-	UPROPERTY(Transient)
-	EExecutionDecision Decision = EExecutionDecision::None;
-
-	UPROPERTY(Transient)
-	EExecutionRelationship Relationship = EExecutionRelationship::None;
-
-public:
-	bool IsAccepted() const
-	{
-		return Decision == EExecutionDecision::Accept;
-	}
-};
-
-USTRUCT(BlueprintType)
-struct FExecutionInterventionQuery
-{
-	GENERATED_BODY()
-
-public:
-	UPROPERTY(Transient)
-	FExecutionSnapshot Snapshot = FExecutionSnapshot();
-
-	UPROPERTY(Transient)
-	FExecutionParticipant IncomingPart = FExecutionParticipant();
-
-	UPROPERTY(Transient)
-	FExecutionParticipant ActivePart = FExecutionParticipant();
-
-	UPROPERTY(Transient)
-	EExecutionStopReason StopReason = EExecutionStopReason::None;
-
-public:
-	bool IsValidMinimal() const
-	{
-		return IncomingPart.IsValidMinimal()
-			&& ActivePart.IsValidMinimal()
-			&& StopReason != EExecutionStopReason::None
-			&& StopReason != EExecutionStopReason::Max;
-	}
-
-};
-
-USTRUCT(BlueprintType)
-struct FExecutionInterventionDirective
-{
-	GENERATED_BODY()
-
-public:
-	UPROPERTY(Transient)
-	bool bRequested = false;
-
-	UPROPERTY(Transient)
-	EExecutionStopSource StopSource = EExecutionStopSource::None;
-
-	UPROPERTY(Transient)
-	EExecutionDomain SourceDomain = EExecutionDomain::None;
-
-	UPROPERTY(Transient)
-	EExecutionDomain TargetDomain = EExecutionDomain::None;
-
-	UPROPERTY(Transient)
-	EExecutionStopReason StopReason = EExecutionStopReason::None;
-
-	UPROPERTY(Transient)
-	EExecutionAfterStopAction AfterStopAction = EExecutionAfterStopAction::None;
-
-	UPROPERTY(Transient)
-	FExecutionParticipant IncomingPart = FExecutionParticipant();
-
-	UPROPERTY(Transient)
-	FExecutionParticipant ActivePart = FExecutionParticipant();
-
-public:
-	bool IsRequested() const
-	{
-		return bRequested;
-	}
-
-	bool IsValidRequest() const
-	{
-		return bRequested
-			&& StopSource != EExecutionStopSource::None
-			&& StopSource != EExecutionStopSource::Max
-			&& SourceDomain != EExecutionDomain::None
-			&& SourceDomain != EExecutionDomain::Max
-			&& TargetDomain != EExecutionDomain::None
-			&& TargetDomain != EExecutionDomain::Max
-			&& StopReason != EExecutionStopReason::None
-			&& StopReason != EExecutionStopReason::Max
-			&& IncomingPart.IsValidMinimal()
-			&& ActivePart.IsValidMinimal();
-	}
 };
 
 USTRUCT(BlueprintType)
